+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 10.1.3 (Unreleased)
 
+- Add the ability to edit the name and group of a roster contact throught the edit user details modal
 - Add the ability to set roles and affiliations via the MUC occupant modal
 - Fix `isOnlyEmojis is not a function` when using only `@converse/headless`
 - Fix `autojoin` checkbox state in MUC bookmark form